refactor: Service test에 중복 코드 발생이 넘쳐날 낌새가 보여 superclass로 ServiceTestBase 생성
